+++ src/components/RegistrationForm/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Address';
+export * from './Finish';
+export * from './PersonalData';
