--- src/components/registration-form/components/index.ts
@@ -1,3 +1,0 @@
-export * from './address';
-export * from './finish';
-export * from './personal-data';
