--- src/components/registration-form/index.ts
@@ -1,1 +1,0 @@
-export * from './registration-form';
